Moved Paginator to Utils as it's fairly generic now
